requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.5
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.1.2
marshmallow==3.23.1
marshmallow-enum==1.5.1